Implement New Design (#27)
* Update files according to new design elements

This commit updates the HTML, CSS, and JS for the main site to match the
elements in the new design. It also adds "TO-DO" comments for future
tasks.

* Update CSS according to new design

This commit updates the CSS to match the new design. Further consideration
may need to be done in terms of font family and font weight.

* Add and organize graphics

This commit adds the graphics to the site and provides basic styling.
Mobile styling has not been completed yet.

* Remove outdated comments

* Fix banner graphic size

* Fix button text

---------

Co-authored-by: tdianaa <rtoneyd@gmail.com>